app/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langfuse>=3.12.1